Switch to color.RGBA
it's a little extra work when sending data to wasm,
but closer to what the real code will be. It also make
it easier for me to write scale() that will carry over to
the SAMD51 board.

I also fixed the bug in sprite.Render that was causing a
black square to appear on "full" loaders.